requirements.txt
pandas==2.0.3
pyro-ppl==1.8.6
pytz==2023.3
scikit-learn==1.3.0
yfinance==0.2.27
